I refactored the classes in geometry,py into their own modules within a package called geometry.
