+++ .idea/PteraSoftware.iml
@@ -13,7 +13,7 @@
       <excludeFolder url="file://$MODULE_DIR$/.venv310" />
       <excludeFolder url="file://$MODULE_DIR$/.venv311" />
       <excludeFolder url="file://$MODULE_DIR$/.venv312" />
-      <excludeFolder url="file://$MODULE_DIR$/.venv313" />
+      <excludeFolder url="file://$MODULE_DIR$/.venv" />
     </content>
     <orderEntry type="jdk" jdkName="Python 3.13 (PteraSoftware)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
