+++ pterasoftware/geometry/__init__.py
@@ -1,0 +1,24 @@
+"""This package contains the geometry classes.
+
+This package contains the following subpackages:
+    None
+
+This package contains the following directories:
+    None
+
+This package contains the following modules:
+    __init__.py: This module is this package's initialization script.
+
+    airfoil.py: This module contains the Airfoil class.
+
+    airplane.py: This module contains the Airplane class.
+
+    wing.py: This module contains the Wing class.
+
+    wing_cross_section.py: This module contains the WingCrossSection class.
+"""
+
+import pterasoftware.geometry.airfoil
+import pterasoftware.geometry.airplane
+import pterasoftware.geometry.wing
+import pterasoftware.geometry.wing_cross_section
