+++ pterasoftware/geometry/wing.py
@@ -1,0 +1,817 @@
+"""This module contains the Wing class.
+
+This module contains the following classes:
+    Wing: This is a class used to contain wings of an Airplane.
+
+This module contains the following exceptions:
+    None
+
+This module contains the following functions:
+    None
+"""
+
+import numpy as np
+
+from .wing_cross_section import WingCrossSection
+
+# ToDo: Uncomment this line once we uncomment the meshing command.
+# from .. import meshing
+from .. import parameter_validation
+from .. import transformations
+
+
+class Wing:
+    """This is a class used to contain the wings of an Airplane.
+
+    Citation:
+        Adapted from:         geometry.Wing in AeroSandbox
+        Author:               Peter Sharpe
+        Date of Retrieval:    04/24/2020
+
+    This class contains the following public methods:
+
+        generate_mesh: This method generates this Wing's mesh, which finishes the
+        process of preparing the Wing to be used in a simulation. It is called by the
+        Wing's parent Airplane, after it's determined its symmetry type.
+
+        T_pas_G_Cg_to_Wn_Ler: This method defines a property for the passive
+        transformation matrix which maps in homogeneous coordinates from geometry
+        axes relative to the CG point to wing axes relative to the leading edge root
+        point. This is set to None if the Wing's symmetry type hasn't been defined yet.
+
+        T_pas_Wn_Ler_to_G_Cg: This method defines a property for the passive
+        transformation matrix which maps in homogeneous coordinates from wing axes
+        relative to the leading edge root point to geometry axes relative to the CG
+        point. This is set to None if the Wing's symmetry type hasn't been defined yet.
+
+        WnX_G: This method sets a property for the wing axes' first basis vector (in
+        geometry axes).
+
+        WnY_G: This method sets a property for the wing axes' second basis vector (in
+        geometry axes).
+
+        WnZ_G: This method sets a property for the wing axes' third basis vector (in
+        geometry axes).
+
+        projected_area: This method sets a property for the area of the Wing
+        projected onto the plane defined by the wing axes' xy-plane.
+
+        wetted_area: This method sets a property for the Wing's wetted area.
+
+        span: This method sets a property for the Wing's span.
+
+        standard_mean_chord: This method sets a property for the standard mean chord
+        of the Wing.
+
+        mean_aerodynamic_chord: This method sets a property for the mean aerodynamic
+        chord of the Wing.
+
+    This class contains the following class attributes:
+        None
+
+    Subclassing:
+        This class is not meant to be subclassed.
+
+    Every Wing has its axis system, known as wing axes. The user sets the
+    relationship between these axes and geometry axes with the prelimLer_G_Cg and
+    angles_G_to_prelimWn parameters. However, the steps for transforming a vector
+    from geometry axes to wing axes, and the interpretation of the wing axes
+    orientation and position relative to a Wing's geometry, also depend on the
+    parameters symmetric, mirror_only, symmetry_normal_prelimWn,
+    and symmetry_point_prelimWn_prelimLer:
+
+        1. symmetric is False
+
+            A. mirror_only is False
+
+                I. The symmetry plane must be undefined (symmetry_normal_prelimWn and
+                symmetry_point_prelimWn_prelimLer must be None)
+
+                    Type 1 Symmetry:
+
+                    - prelimLer_G_Cg is the final location of the leading edge of
+                    this Wing's root WingCrossSection, as defined in geometry axes.
+
+                    - prelimLer_G_Cg is also the final location of the origin of this
+                    Wing's axes, as defined in geometry axes.
+
+                    - Translation by prelimLer_G_Cg followed by rotations by
+                    angles_G_to_prelimWn fully define this Wing's axes with respect
+                    to the geometry axes. The wing axes will also retain the
+                    handedness of the geometry axes.
+
+            B. mirror_only is True
+
+                I. The symmetry plane is coincident with this Wing's axes' xz-plane
+
+                    Type 2 Symmetry:
+
+                    - prelimLer_G_Cg is the final location of the leading edge of
+                    this Wing's root WingCrossSection, as defined in geometry axes.
+
+                    - prelimLer_G_Cg is also the final location of the origin of this
+                    Wing's axes, as defined in geometry axes.
+
+                    - Translation by prelimLer_G_Cg followed by rotations by
+                    angles_G_to_prelimWn does not fully define the orientation of
+                    this Wing's axes with respect to the geometry axes. After
+                    translation and rotation, the coordinate system also needs to be
+                    reflected across the symmetry plane, which will flip the wing
+                    axes' handedness to be opposite that of geometry axes.
+
+                II. The symmetry plane is not coincident with this Wing's axes'
+                xz-plane
+
+                    Type 3 Symmetry:
+
+                    - prelimLer_G_Cg is not final location of the leading edge of
+                    this Wing's root WingCrossSection, as defined in geometry axes.
+
+                    - prelimLer_G_Cg is not the final location of the origin of this
+                    Wing's axes, as defined in geometry axes.
+
+                    - Translation by prelimLer_G_Cg followed by rotations by
+                    angles_G_to_prelimWn does not fully define orientation of this
+                    Wing's axes with respect to the geometry axes. After translation
+                    and rotation, the coordinate system also needs to be reflected
+                    across the symmetry plane, which will flip the wing axes'
+                    handedness to be opposite that of geometry axes.
+
+        2. symmetric is True
+
+            A. mirror_only must be False
+
+                I. the symmetry plane is coincident with this Wing's axes' xz-plane
+
+                    Type 4 Symmetry:
+
+                    - prelimLer_G_Cg is the final location of the leading edge of
+                    this Wing's root WingCrossSection, as defined in geometry axes.
+                    However, while the root WingCrossSection is the still the first
+                    item in the wing_cross_sections list, when meshed, panels will
+                    extend from the root in both the +y and -y wing axis directions.
+                    The length of the wing_cross_sections list remains unchanged.
+
+                    - prelimLer_G_Cg is also the final location of the origin of this
+                    Wing's axes, as defined in geometry axes.
+
+                    - Translation by prelimLer_G_Cg followed by rotations by
+                    angles_G_to_prelimWn fully define this Wing's axes with respect
+                    to the geometry axes. The wing axes will also retain the
+                    handedness of the geometry axes.
+
+                II. the symmetry plane is not coincident with this Wing's axes' xz-plane
+
+                    Type 5 Symmetry:
+
+                    - This Wing's Airplane will set this Wing's symmetric parameter
+                    to False, its mirror_only parameter to False,
+                    its symmetry_normal_prelimWn parameter to None and its
+                    symmetry_point_prelimWn_prelimLer parameter to None. These
+                    changes turn this Wing into a "Scenario 1 Wing."
+
+                    - The Airplane will also create a new Wing, and add it to its
+                    wings list immediately after this Wing. The new Wing will have
+                    the same name as this Wing, but with the prefix "Reflected ". The
+                    new Wing also will have all the same parameters as this Wing,
+                    except that symmetric will be False and mirror_only will be True,
+                    which means that it will be a "Scenario 3 Wing."
+
+                    - Also, if the control_surface_type is "asymmetric" for any of
+                    this Wing's WingCrossSections, the reflected Wing's corresponding
+                    WingCrossSections will have their control_surface_deflection
+                    values multiplied by negative one.
+    """
+
+    def __init__(
+        self,
+        wing_cross_sections,
+        name="Untitled Wing",
+        prelimLer_G_Cg=(0.0, 0.0, 0.0),
+        angles_G_to_prelimWn=(0.0, 0.0, 0.0),
+        symmetric=False,
+        mirror_only=False,
+        symmetry_normal_prelimWn=None,
+        symmetry_point_prelimWn_prelimLer=None,
+        num_chordwise_panels=8,
+        chordwise_spacing="cosine",
+    ):
+        """This is the initialization method.
+
+        :param wing_cross_sections: list of WingCrossSections
+
+            This is a list of WingCrossSections that represent the wing's cross
+            sections in order from root to tip. It must contain at least two
+            WingCrossSections.
+
+        :param name: str, optional
+
+            This is a sensible name for the Wing. The default is "Untitled Wing".
+
+        :param prelimLer_G_Cg: array-like of 3 numbers, optional
+
+            This is the position [x, y, z] of the origin of this Wing's axes (in
+            geometry axes, relative to the starting point) before any symmetry or
+            mirror has been applied. Can be a tuple, list, or numpy array of numbers
+            (int or float). Values are converted to floats internally. It may differ
+            from the actual position as explained in the class docstring. The units are
+            meters. The default is (0.0, 0.0, 0.0).
+
+        :param angles_G_to_prelimWn: array-like of 3 numbers, optional
+
+            This is the rotation angles [roll, pitch, yaw] in degrees that define the
+            orientation of this Wing's axes relative to the geometry axes before any
+            symmetry or mirror has been applied. Can be a tuple, list, or numpy array
+            of numbers (int or float). Values are converted to floats internally. All
+            angles must be in the range (-90, 90) degrees. Roll is rotation about the
+            x-axis, pitch is rotation about the y-axis, and yaw is rotation about the
+            z-axis. Rotations are intrinsic, and proceed in the z-y'-x'' order
+            conventional for Euler angles. It may differ from the actual position as
+            explained in the class docstring. The units are degrees. The default is (
+            0.0, 0.0, 0.0).
+
+        :param symmetric: bool, optional
+
+            Set this to True if the Wing's geometry should be mirrored across the
+            symmetry plane while retaining the non-mirrored side. If mirror_only is
+            True, symmetric must be False. If symmetric is true, then neither
+            symmetry_normal_prelimWn nor symmetry_point_prelimWn_prelimLer can be
+            None. If the symmetry plane is coincident with this Wing's wing axes'
+            xz-plane, the mirrored and non-mirrored geometry will be meshed as a
+            single wing. If not, this Wing's Airplane will automatically create
+            another Wing with the mirrored geometry, modify both Wings' parameters,
+            and add the reflected Wing to its list of wings immediately following
+            this one. For more details on how that process, and how this parameter
+            interacts with symmetry_normal_prelimWn,
+            symmetry_point_prelimWn_prelimLer, and mirror_only, see the class
+            docstring. The default is False.
+
+        :param mirror_only: bool, optional
+
+            Set this to True if the Wing's geometry should be reflected about the
+            symmetry plane without retaining the non-reflected geometry. If symmetric
+            is True, mirror_only must be False. If mirror_only is true, then neither
+            symmetry_normal_prelimWn nor symmetry_point_prelimWn_prelimLer can be
+            None. For more details on how this parameter interacts with
+            symmetry_normal_prelimWn, symmetry_point_prelimWn_prelimLer,
+            and symmetric, see the class docstring. The default is False.
+
+        :param symmetry_normal_prelimWn: array-like of 3 numbers or None, optional
+
+            The unit normal vector (in preliminary wing axes) that, together with
+            symmetry_point_prelimWn_prelimLer, defines the plane used for symmetry or
+            mirroring. Can be a tuple, list, or numpy array of numbers (int or
+            float), or None. Values are converted to floats and normalized
+            internally. Note that reversing the normal direction (using the
+            antiparallel vector) defines the same plane and produces the same result.
+            This value must be None if both symmetric and mirror_only are False,
+            and cannot be None if either are True. For more details on how this
+            parameter interacts with symmetry_point_prelimWn_prelimLer, symmetric,
+            and mirror_only, see the class docstring. The default is None.
+
+        :param symmetry_point_prelimWn_prelimLer: array-like of 3 numbers or None,
+        optional
+
+            A point [x, y, z] (in preliminary wing axes, relative to the preliminary
+            leading edge root point) that, along with symmetry_normal_prelimWn,
+            defines the location of the plane about which symmetry or mirroring is
+            applied. Can be a list, tuple, or numpy array of numbers (int or float),
+            or None. Values are converted to floats internally. This value must be
+            None if both symmetric and mirror_only are False, and cannot be None if
+            either are True. For more details on how this parameter interacts with
+            symmetry_normal_prelimWn, symmetric, and mirror_only, see the class
+            docstring. The units are meters. The default is None.
+
+        :param num_chordwise_panels: int, optional
+
+            This is the number of chordwise panels to be used on this wing,
+            which must be set to a positive integer. The default is 8.
+
+        :param chordwise_spacing: str, optional
+
+            This is the type of spacing between the wing's chordwise panels. It can
+            be "cosine" or "uniform". Using cosine spacing is highly recommended for
+            steady simulations and uniform spacing is highly recommended for unsteady
+            simulations. The default is "cosine".
+
+        """
+        # Validate wing_cross_sections.
+        wing_cross_sections = parameter_validation.validate_non_empty_list(
+            wing_cross_sections, "wing_cross_sections"
+        )
+        num_wing_cross_sections = len(wing_cross_sections)
+        if num_wing_cross_sections < 2:
+            raise ValueError("wing_cross_sections must contain at least two elements.")
+        for wing_cross_section_id, wing_cross_section in enumerate(wing_cross_sections):
+            if not isinstance(wing_cross_section, WingCrossSection):
+                raise TypeError(
+                    "Every element in wing_cross_sections must be a WingCrossSection."
+                )
+            if wing_cross_section_id == 0:
+                # Validate root WingCrossSection constraints.
+                wing_cross_section.validate_root_constraints()
+            elif wing_cross_section_id == num_wing_cross_sections - 1:
+                # Validate tip WingCrossSection constraints.
+                wing_cross_section.validate_tip_constraints()
+            # Set the validated flag for this WingCrossSection.
+            wing_cross_section.validated = True
+        self.wing_cross_sections = wing_cross_sections
+
+        # Validate name and prelimLer_G_Cg.
+        self.name = parameter_validation.validate_string(name, "name")
+        self.prelimLer_G_Cg = parameter_validation.validate_3d_vector_float(
+            prelimLer_G_Cg, "prelimLer_G_Cg"
+        )
+
+        # Validate angles_G_to_prelimWn.
+        angles_G_to_prelimWn = parameter_validation.validate_3d_vector_float(
+            angles_G_to_prelimWn, "angles_G_to_prelimWn"
+        )
+        if not np.all((-90.0 < angles_G_to_prelimWn) & (angles_G_to_prelimWn < 90.0)):
+            raise ValueError(
+                "All elements of angles_G_to_prelimWn must lie in the range (-90, 90) degrees."
+            )
+        self.angles_G_to_prelimWn = angles_G_to_prelimWn
+
+        # Validate symmetric and mirror_only.
+        symmetric = parameter_validation.validate_boolean(symmetric, "symmetric")
+        mirror_only = parameter_validation.validate_boolean(mirror_only, "mirror_only")
+        if symmetric and mirror_only:
+            raise ValueError("symmetric and mirror_only cannot both be True.")
+        self.symmetric = symmetric
+        self.mirror_only = mirror_only
+
+        # Validate symmetry_normal_prelimWn and symmetry_point_prelimWn_prelimLer.
+        if self.symmetric or self.mirror_only:
+            if symmetry_normal_prelimWn is None:
+                raise ValueError(
+                    "symmetry_normal_prelimWn cannot be None when symmetric or mirror_only is True."
+                )
+            symmetry_normal_prelimWn = (
+                parameter_validation.validate_3d_unit_vector_norm_float(
+                    symmetry_normal_prelimWn, "symmetry_normal_prelimWn"
+                )
+            )
+            if symmetry_point_prelimWn_prelimLer is None:
+                raise ValueError(
+                    "symmetry_point_prelimWn_prelimLer cannot be None when symmetric or mirror_only is True."
+                )
+            symmetry_point_prelimWn_prelimLer = (
+                parameter_validation.validate_3d_vector_float(
+                    symmetry_point_prelimWn_prelimLer,
+                    "symmetry_point_prelimWn_prelimLer",
+                )
+            )
+        else:
+            if symmetry_normal_prelimWn is not None:
+                raise ValueError(
+                    "symmetry_normal_prelimWn must be None when both symmetric and mirror_only are False."
+                )
+            if symmetry_point_prelimWn_prelimLer is not None:
+                raise ValueError(
+                    "symmetry_point_prelimWn_prelimLer must be None when both symmetric and mirror_only are False."
+                )
+        self.symmetry_normal_prelimWn = symmetry_normal_prelimWn
+        self.symmetry_point_prelimWn_prelimLer = symmetry_point_prelimWn_prelimLer
+
+        # Validate num_chordwise_panels and chordwise_spacing.
+        self.num_chordwise_panels = parameter_validation.validate_positive_scalar_int(
+            num_chordwise_panels, "num_chordwise_panels"
+        )
+        if chordwise_spacing not in ["cosine", "uniform"]:
+            raise ValueError('chordwise_spacing must be "cosine" or "uniform".')
+        self.chordwise_spacing = chordwise_spacing
+
+        # These attributes will be initialized or populated once this Wing's parent
+        # Airplane calls generate_mesh.
+        self.symmetry_type = None
+        self.num_spanwise_panels = None
+        self.num_panels = None
+        self.wake_ring_vortex_vertices = None
+        self.wake_ring_vortices = None
+        self.panels = None
+
+    def generate_mesh(self, symmetry_type):
+        """This method generates this Wing's mesh, which finishes the process of
+        preparing the Wing to be used in a simulation. It is called by the Wing's
+        parent Airplane, after it's determined its symmetry type.
+
+        :param symmetry_type:
+        :return:
+        """
+        # Validate and apply symmetry_type. 5 isn't a valid symmetry type, because
+        # the parent Airplane should have modified a Wing that initially had type 5
+        # symmetry to have type 1 symmetry, and then made a new reflected Wing with
+        # type 3 symmetry.
+        symmetry_type = parameter_validation.validate_scalar_int(
+            symmetry_type, "symmetry_type"
+        )
+        valid_symmetry_types = [1, 2, 3, 4]
+        if symmetry_type not in valid_symmetry_types:
+            raise ValueError(f"symmetry_type must be one of {valid_symmetry_types}")
+        self.symmetry_type = symmetry_type
+
+        # Find the number of spanwise panels on the wing by adding each cross
+        # section's number of spanwise panels. Exclude the last cross section's
+        # number of spanwise panels as this is irrelevant. If the wing has type 4
+        # symmetry multiply the summation by two.
+        self.num_spanwise_panels = 0
+        for wing_cross_section in self.wing_cross_sections[:-1]:
+            self.num_spanwise_panels += wing_cross_section.num_spanwise_panels
+        if self.symmetry_type == 4:
+            self.num_spanwise_panels *= 2
+
+        # Calculate the number of panels on this wing.
+        self.num_panels = self.num_spanwise_panels * self.num_chordwise_panels
+
+        # Initialize an empty array to hold this wing's wake ring vortices and its
+        # wake ring vortex vertices.
+        self.wake_ring_vortex_vertices = np.empty(
+            (0, self.num_spanwise_panels + 1, 3), dtype=float
+        )
+        self.wake_ring_vortices = np.zeros((0, self.num_spanwise_panels), dtype=object)
+
+        # Generate the wing's mesh, which populates the Panels attribute.
+        # ToDo: Uncomment this once meshing has been fixed for the refactored
+        #  geometry definitions.
+        # meshing.mesh_wing(self)
+
+    @property
+    def T_pas_G_Cg_to_Wn_Ler(self):
+        """This method defines a property for the passive transformation matrix which
+        maps in homogeneous coordinates from geometry axes relative to the CG point
+        to wing axes relative to the leading edge root point. This is set to None if
+        the Wing's symmetry type hasn't been defined yet.
+
+        :return: (4,4) ndarray of floats or None
+            4x4 transformation matrix or None in cases where the Wing's symmetry type
+            hasn't been defined yet.
+        """
+        # If the Wing's symmetry type hasn't been set yet, return None to avoid
+        # incorrect symmetry handling.
+        if self.symmetry_type is None:
+            return None
+
+        # Step 1: Create T_trans_pas_G_Cg_to_G_prelimLer, which maps in homogenous
+        # coordinates from geometry axes relative to the CG point to geometry axes
+        # relative to the preliminary leading edge root point. This is the
+        # translation step.
+        T_trans_pas_G_Cg_to_G_prelimLer = transformations.generate_T_trans(
+            self.prelimLer_G_Cg, passive=True
+        )
+
+        # Step 2: Create T_rot_pas_G_to_prelimWn, which maps in homogeneous
+        # coordinates from geometry axes to preliminary wing axes. This is the
+        # rotation step.
+        T_rot_pas_G_to_prelimWn = transformations.generate_T_rot(
+            transformations.generate_R(
+                self.angles_G_to_prelimWn, passive=True, intrinsic=True, order="321"
+            )
+        )
+
+        # Step 3: Create T_reflect_pas_prelimWn_prelimLer_to_Wn_Ler, which maps from
+        # which maps in homogeneous coordinates from preliminary wing axes relative
+        # to the preliminary leading edge root point to wing axes relative to the
+        # leading edge root point. This is the reflection step.
+        if (
+            self.symmetry_normal_prelimWn is not None
+            and self.symmetry_point_prelimWn_prelimLer is not None
+        ):
+            T_reflect_pas_prelimWn_prelimLer_to_Wn_Ler = (
+                transformations.generate_T_reflect(
+                    self.symmetry_point_prelimWn_prelimLer,
+                    self.symmetry_normal_prelimWn,
+                    passive=True,
+                )
+            )
+        else:
+            T_reflect_pas_prelimWn_prelimLer_to_Wn_Ler = np.eye(4, dtype=float)
+
+        return (
+            T_reflect_pas_prelimWn_prelimLer_to_Wn_Ler
+            @ T_rot_pas_G_to_prelimWn
+            @ T_trans_pas_G_Cg_to_G_prelimLer
+        )
+
+    @property
+    def T_pas_Wn_Ler_to_G_Cg(self):
+        """This method defines a property for the passive transformation matrix which
+        maps in homogeneous coordinates from wing axes relative to the leading edge
+        root point to geometry axe
+        s relative to the CG point. This is set to None if
+        the Wing's symmetry type hasn't been defined yet.
+
+        :return: (4,4) ndarray of floats or None
+            4x4 transformation matrix or None in cases where the Wing's symmetry type
+            hasn't been defined yet.
+        """
+        # If the Wing's symmetry type hasn't been set yet, return None to avoid
+        # incorrect symmetry handling.
+        if self.symmetry_type is None:
+            return None
+
+        return np.linalg.inv(self.T_pas_G_Cg_to_Wn_Ler)
+
+    @property
+    def WnX_G(self):
+        """This method sets a property for the wing axes' first basis vector (in
+        geometry axes).
+
+        :return: (3,) ndarray of floats or None
+            This is the wing axes' first basis vector (in geometry axes) or None in
+            cases where the Wing's symmetry type hasn't been defined yet.
+        """
+        # If the Wing's symmetry type hasn't been set yet, return None to avoid
+        # incorrect symmetry handling.
+        if self.symmetry_type is None:
+            return None
+
+        WnX_Wn = np.array([1.0, 0.0, 0.0])
+
+        WnXHomog_Wn = transformations.generate_homog(WnX_Wn, has_point=False)
+
+        WnXHomog_G = self.T_pas_Wn_Ler_to_G_Cg @ WnXHomog_Wn
+
+        return WnXHomog_G[:3]
+
+    @property
+    def WnY_G(self):
+        """This method sets a property for the wing axes' second basis vector (in
+        geometry axes).
+
+        :return: (3,) ndarray of floats or None
+            This is the wing axes' second basis vector (in geometry axes) or None in
+            cases where the Wing's symmetry type hasn't been defined yet.
+        """
+        # If the Wing's symmetry type hasn't been set yet, return None to avoid
+        # incorrect symmetry handling.
+        if self.symmetry_type is None:
+            return None
+
+        WnY_Wn = np.array([0.0, 1.0, 0.0])
+
+        WnYHomog_Wn = transformations.generate_homog(WnY_Wn, has_point=False)
+
+        WnYHomog_G = self.T_pas_Wn_Ler_to_G_Cg @ WnYHomog_Wn
+
+        return WnYHomog_G[:3]
+
+    @property
+    def WnZ_G(self):
+        """This method sets a property for the wing axes' third basis vector (in
+        geometry axes).
+
+        :return: (3,) ndarray of floats or None
+            This is the wing axes' third basis vector (in geometry axes) or None in
+            cases where the Wing's symmetry type hasn't been defined yet.
+        """
+        # If the Wing's symmetry type hasn't been set yet, return None to avoid
+        # incorrect symmetry handling.
+        if self.symmetry_type is None:
+            return None
+
+        WnZ_Wn = np.array([0.0, 0.0, 1.0])
+
+        WnZHomog_Wn = transformations.generate_homog(WnZ_Wn, has_point=False)
+
+        WnZHomog_G = self.T_pas_Wn_Ler_to_G_Cg @ WnZHomog_Wn
+
+        return WnZHomog_G[:3]
+
+    @property
+    def projected_area(self):
+        """This method sets a property for the area of the Wing projected onto the
+        plane defined by the wing axes' xy-plane.
+
+        If the Wing is symmetric and continuous, the area of the mirrored half is
+        included.
+
+        :return projected_area: float or None
+            This attribute is the projected area of the wing. It has units of square
+            meters. If the Wing hasn't been meshed yet, None is returned instead.
+        """
+        # Return None if the Wing hasn't been meshed yet.
+        if self.panels is None:
+            return None
+
+        projected_area = 0
+
+        # Iterate through the chordwise and spanwise indices of the panels and add
+        # their area to the total projected area.
+        for chordwise_location in range(self.num_chordwise_panels):
+            for spanwise_location in range(self.num_spanwise_panels):
+                projected_area += self.panels[
+                    chordwise_location, spanwise_location
+                ].calculate_projected_area(self.WnZ_G)
+
+        return projected_area
+
+    @property
+    def wetted_area(self):
+        """This method sets a property for the Wing's wetted area.
+
+        If the Wing is symmetric and continuous, the area of the mirrored half is
+        included.
+
+        :return wetted_area: float or None
+            This attribute is the wetted area of the wing. It has units of square
+            meters. If the Wing hasn't been meshed yet, None is returned instead.
+        """
+        # Return None if the Wing hasn't been meshed yet.
+        if self.panels is None:
+            return None
+
+        wetted_area = 0
+
+        # Iterate through the chordwise and spanwise indices of the panels and add
+        # their area to the total wetted area.
+        for chordwise_location in range(self.num_chordwise_panels):
+            for spanwise_location in range(self.num_spanwise_panels):
+                wetted_area += self.panels[chordwise_location, spanwise_location].area
+
+        return wetted_area
+
+    @property
+    def span(self):
+        """This method sets a property for the Wing's span.
+
+        The span is found by first finding vector connecting the leading edges of the
+        root and tip WingCrossSections. Then, this vector is projected onto the wing
+        axes' third basis vector. The span is defined as the magnitude of this
+        projection.
+
+        If the Wing is symmetric and continuous, this method includes the span of
+        the mirrored half.
+
+        :return span: float or None
+            This is the Wing's span. It has units of meters. None is returned in
+            cases where the Wing's symmetry type hasn't been defined yet
+        """
+        # If the Wing's symmetry type hasn't been set yet, return None to avoid
+        # incorrect symmetry handling.
+        if self.symmetry_type is None:
+            return None
+
+        # Calculate the tip WingCrossSection's leading point position by chaining
+        # transformations through all WingCrossSections from root to tip.
+
+        # Start at root leading point (wing axes origin)
+        current_position_Wn_Ler = np.array([0.0, 0.0, 0.0])
+
+        # For each subsequent WingCrossSection, accumulate its displacement
+        for i in range(1, len(self.wing_cross_sections)):
+            wing_cross_section = self.wing_cross_sections[i]
+
+            # Get displacement in parent axes (WingCrossSection[i-1]'s axes)
+            displacement_parent = wing_cross_section.Lp_Wcsp_Lpp
+
+            # Compute transformation from parent axes to wing axes
+            T_parent_to_wing = np.eye(4, dtype=float)
+            for j in range(i - 1, 0, -1):  # Go backwards from parent to root
+                T_parent_to_wing = (
+                    self.wing_cross_sections[j].T_pas_Wcs_Lp_to_Wcsp_Lpp
+                    @ T_parent_to_wing
+                )
+
+            # Transform displacement from parent axes to wing axes
+            displacementHomog_parent = transformations.generate_homog(
+                displacement_parent, has_point=False
+            )
+            displacementHomog_wing = T_parent_to_wing @ displacementHomog_parent
+
+            # Add to accumulated position
+            current_position_Wn_Ler += displacementHomog_wing[:3]
+
+        # Now we have tip position in wing axes relative to Ler
+        tipLp_Wn_Ler = current_position_Wn_Ler
+
+        # Project the tip position onto the wing axes' y-direction (spanwise direction)
+        projected_tipLp_Wn_Ler = np.dot(
+            tipLp_Wn_Ler, np.array([0.0, 1.0, 0.0])
+        ) * np.array([0.0, 1.0, 0.0])
+
+        span = np.linalg.norm(projected_tipLp_Wn_Ler)
+
+        # If the wing is symmetric and continuous, multiply the span by two.
+        if self.symmetry_type == 4:
+            span *= 2
+
+        return span
+
+    @property
+    def standard_mean_chord(self):
+        """This method sets a property for the standard mean chord of the Wing.
+
+        The standard mean chord is defined as the projected area divided by the span.
+        See their respective methods for the definitions of span and projected area.
+
+        :return: float or None
+            This is the standard mean chord of the Wing. It has units of meters. None
+            is returned in cases where the Wing's symmetry type hasn't been defined yet.
+        """
+        # If the Wing's symmetry type hasn't been set yet, return None to avoid
+        # incorrect symmetry handling.
+        if self.symmetry_type is None:
+            return None
+
+        return self.projected_area / self.span
+
+    @property
+    def mean_aerodynamic_chord(self):
+        """This method sets a property for the mean aerodynamic chord of the Wing.
+
+        :return: float or None
+            This is the mean aerodynamic chord of the Wing. It has units of meters.
+            None is returned in cases where the Wing's symmetry type hasn't been
+            defined yet.
+        """
+        # If the Wing's symmetry type hasn't been set yet, return None to avoid
+        # incorrect symmetry handling.
+        if self.symmetry_type is None:
+            return None
+
+        # This method is based on the equation for the mean aerodynamic chord of a
+        # wing, which can be found here: https://en.wikipedia.org/wiki/Chord_(
+        # aeronautics)#Mean_aerodynamic_chord. This equation integrates the squared
+        # chord from the Wing's center to the Wing's tip. We will perform this
+        # integral piecewise for each section of the Wing.
+        integral = 0
+
+        # Iterate through the WingCrossSections to add the contribution of their
+        # corresponding Wing section to the piecewise integral.
+        for wing_cross_section_id, wing_cross_section in enumerate(
+            self.wing_cross_sections[:-1]
+        ):
+            next_wing_cross_section = self.wing_cross_sections[
+                wing_cross_section_id + 1
+            ]
+
+            chord = wing_cross_section.chord
+            next_chord = next_wing_cross_section.chord
+
+            # Find this section's span by calculating the positions of both
+            # WingCrossSections in wing axes, then finding the distance between them.
+
+            # Calculate current WingCrossSection's position in wing axes
+            current_position_Wn_Ler = np.array([0.0, 0.0, 0.0])
+            for k in range(1, wing_cross_section_id + 1):
+                wcs = self.wing_cross_sections[k]
+                displacement_parent = wcs.Lp_Wcsp_Lpp
+
+                # Transform from parent axes to wing axes
+                T_parent_to_wing = np.eye(4, dtype=float)
+                for j in range(k - 1, 0, -1):
+                    T_parent_to_wing = (
+                        self.wing_cross_sections[j].T_pas_Wcs_Lp_to_Wcsp_Lpp
+                        @ T_parent_to_wing
+                    )
+
+                displacementHomog_parent = transformations.generate_homog(
+                    displacement_parent, has_point=False
+                )
+                displacementHomog_wing = T_parent_to_wing @ displacementHomog_parent
+                current_position_Wn_Ler += displacementHomog_wing[:3]
+
+            # Calculate next WingCrossSection's position in wing axes
+            next_position_Wn_Ler = np.array([0.0, 0.0, 0.0])
+            for k in range(1, wing_cross_section_id + 2):
+                wcs = self.wing_cross_sections[k]
+                displacement_parent = wcs.Lp_Wcsp_Lpp
+
+                # Transform from parent axes to wing axes
+                T_parent_to_wing = np.eye(4, dtype=float)
+                for j in range(k - 1, 0, -1):
+                    T_parent_to_wing = (
+                        self.wing_cross_sections[j].T_pas_Wcs_Lp_to_Wcsp_Lpp
+                        @ T_parent_to_wing
+                    )
+
+                displacementHomog_parent = transformations.generate_homog(
+                    displacement_parent, has_point=False
+                )
+                displacementHomog_wing = T_parent_to_wing @ displacementHomog_parent
+                next_position_Wn_Ler += displacementHomog_wing[:3]
+
+            # Find the section vector and project it onto spanwise direction
+            section_vector_Wn = next_position_Wn_Ler - current_position_Wn_Ler
+
+            # Project section vector onto spanwise direction (wing axes y-direction)
+            projected_section_vector = np.dot(
+                section_vector_Wn, np.array([0.0, 1.0, 0.0])
+            ) * np.array([0.0, 1.0, 0.0])
+
+            section_span = np.linalg.norm(projected_section_vector)
+
+            # Each Wing section is, by definition, trapezoidal (at least when
+            # projected on to the wing's projection plane). For a trapezoid,
+            # the integral from the cited equation can be shown to evaluate to the
+            # following.
+            integral += (
+                section_span * (chord**2 + chord * next_chord + next_chord**2) / 3
+            )
+
+        # Multiply the integral's value by the coefficients from the cited equation.
+        # Double if the wing is symmetric and continuous.
+        if self.symmetry_type == 4:
+            return 2 * integral / self.projected_area
+        return integral / self.projected_area
